@@ -1,7 +1,7 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-/* import './index.css';
-import App from './App'; */
 import { App } from './App';
+import { createRoot } from 'react-dom/client';
 
-ReactDOM.render(<App/>, document.getElementById('root'));
+const domNode = document.getElementById('root');
+const root = createRoot(domNode);
+root.render(<App/>);
